refactor(jobs): supprime les anciens composants de suivi de candidatures
- Retire `JobTrackFormPageComponent` et `JobTrackListComponent` ainsi que leurs dépendances associées
- Nettoie les imports inutilisés et réorganise les fichiers liés pour une structure plus cohérente
- Prépare l'application à l'intégration d'une nouvelle solution de gestion des candidatures
